+++ Aplib.Core.Tests/Collections/ExposedQueueTests.cs
@@ -1,3 +1,4 @@
+using Aplib.Core.Collections;
 using System;
 using System.Collections.Generic;
